requirements.txt
python-dotenv
langchain
langchain-community
langchain-cohere
langchain-groq
cohere
groq
requests
beautifulsoup4
gradio
